[Refactor] Member page 컴포넌트 분리
